20180316 shenlin 数据提取部分：进行基于xml和java的二次提取，获取bugquery文件和corpus文件供后续计算使用。 以VSM为例实现bug报告和源代码之间的相似度计算，并输出文件（未按相似度高低排序）。

+++ v0/src/com/sl/v0/buglocation/models/MyDoubleDictionary.java
@@ -7,7 +7,7 @@ public class MyDoubleDictionary extends HashMap<String, Double>{
 	{
 		if (this.containsKey(term))
 		{
-			this.put(term, this.get(term) + 1);
+			this.put(term, (Double)(this.get(term) + 1));
 		}
 		else
 		{
